+++ checkpoint/orbax/checkpoint/_src/handlers/composite_checkpoint_handler_test.py
@@ -1037,7 +1037,6 @@ class CompositeCheckpointHandlerTest(parameterized.TestCase):
     state_handler.close.assert_called_once()
     metadata_handler.close.assert_called_once()
 
-
   def test_items_exist_final(self):
     handler = CompositeCheckpointHandler('state', 'metadata')
     state = {'a': 1, 'b': 2}