+++ checkpoint/orbax/checkpoint/experimental/v1/_src/metadata/loading.py
@@ -13,11 +13,11 @@
 # limitations under the License.
 
 """Functions for loading metadata from a checkpoint."""
-
-from orbax.checkpoint import handlers
-from orbax.checkpoint._src.checkpointers import checkpointer as checkpointer_lib
+from typing import Any
+from etils import epath
 from orbax.checkpoint.experimental.v1._src.context import context as context_lib
 from orbax.checkpoint.experimental.v1._src.handlers import pytree_handler
+from orbax.checkpoint.experimental.v1._src.loading import loading
 from orbax.checkpoint.experimental.v1._src.metadata import types as metadata_types
 from orbax.checkpoint.experimental.v1._src.path import format_utils
 from orbax.checkpoint.experimental.v1._src.path import types as path_types
@@ -30,21 +30,97 @@ PYTREE_CHECKPOINTABLE_KEY = format_utils.PYTREE_CHECKPOINTABLE_KEY
 def pytree_metadata(
     path: path_types.PathLike,
 ) -> CheckpointMetadata[PyTreeMetadata]:
-  """Loads the PyTree metadata from a checkpoint."""
+  """Loads the PyTree metadata from a checkpoint.
+
+  This function retrieves metadata for a PyTree checkpoint, returning an
+  object of type `CheckpointMetadata[PyTreeMetadata]`. Please see documentation
+  on this class for further details.
+
+  In short, the returned object contains a `metadata` attribute (among other
+  attributes like timestamps), which is an instance of `PyTreeMetadata`. The
+  `PyTreeMetadata` describes information specific to the PyTree itself. The most
+  important such property is the PyTree structure, which is a tree structure
+  matching the structure of the checkpointed PyTree, with leaf metadata objects
+  describing each leaf.
+
+  For example::
+
+    metadata = ocp.pytree_metadata(path)  # CheckpointMetadata[PyTreeMetadata]
+    metadata.metadata # PyTreeMetadata
+    metadata.init_timestamp_nsecs  # Checkpoint creation timestamp.
+
+    metadata.metadata  # PyTree structure.
+
+  The metadata can then be used to inform checkpoint loading. For example::
+
+    metadata = ocp.pytree_metadata(path)
+    restored = ocp.load_pytree(path, metadata)
+
+    # Load with altered properties.
+    def _get_abstract_array(arr):
+      # Assumes all checkpoint leaves are array types.
+      new_dtype = ...
+      new_sharding = ...
+      return jax.ShapeDtypeStruct(arr.shape, new_dtype, sharding=new_sharding)
+
+    metadata = dataclasses.replace(metadata,
+          metadata=jax.tree.map(_get_abstract_array, metadata.metadata)
+    )
+    ocp.load_pytree(path, metadata)
+
+  Args:
+    path: The path to the checkpoint.
+
+  Returns:
+    A `CheckpointMetadata[PyTreeMetadata]` object.
+  """
   format_utils.validate_pytree_checkpoint(path)
-  context = context_lib.get_context()
-  registry = handlers.create_default_handler_registry(
-      **{PYTREE_CHECKPOINTABLE_KEY: pytree_handler.create_v0_handler(context)}
-  )
-  with checkpointer_lib.Checkpointer(
-      handlers.CompositeCheckpointHandler(handler_registry=registry)
-  ) as ckptr:
-    metadata = ckptr.metadata(path)
+  with pytree_handler.pytree_handler_context():
+    metadata = checkpointables_metadata(path)
     return CheckpointMetadata[PyTreeMetadata](
-        metadata=PyTreeMetadata(
-            pytree=metadata.item_metadata[PYTREE_CHECKPOINTABLE_KEY].tree,
-        ),
+        metadata=metadata.metadata[PYTREE_CHECKPOINTABLE_KEY],
         init_timestamp_nsecs=metadata.init_timestamp_nsecs,
         commit_timestamp_nsecs=metadata.commit_timestamp_nsecs,
         custom_metadata=metadata.custom_metadata,
     )
+
+
+def checkpointables_metadata(
+    path: path_types.PathLike,
+) -> CheckpointMetadata[dict[str, Any]]:
+  """Loads all checkpointables metadata from a checkpoint.
+
+  This function is a more general version of `pytree_metadata`. The same
+  `CheckpointMetadata` object is returned (with properties like
+  `init_timestamp_nsecs` as shown above), but the type of the core `metadata`
+  property is a dictionary, mapping checkpointable names to their metadata. This
+  mirrors the return value of `load_checkpointables`, which similarly returns a
+  dictionary mapping checkpointable names to their loaded values.
+
+  For example::
+
+    ocp.save_checkpointables(path, {
+        'foo': Foo(),
+        'bar': Bar(),
+    })
+    metadata = ocp.checkpointables_metadata(path)
+    metadata.metadata  # {'foo': AbstractFoo(), 'bar': AbstractBar()}
+
+  Args:
+    path: The path to the checkpoint.
+
+  Returns:
+    A `CheckpointMetadata[dict[str, Any]]` object.
+  """
+  path = epath.Path(path)
+  checkpointer, _ = loading.get_v0_checkpointer_and_args(
+      path, None, context=context_lib.get_context()
+  )
+  metadata = checkpointer.metadata(path)
+  item_metadata = {k: v for k, v in metadata.item_metadata.items()}
+  return CheckpointMetadata[dict[str, Any]](
+      metadata=item_metadata,
+      init_timestamp_nsecs=metadata.init_timestamp_nsecs,
+      commit_timestamp_nsecs=metadata.commit_timestamp_nsecs,
+      custom_metadata=metadata.custom_metadata,
+  )