+++ checkpoint/orbax/checkpoint/experimental/v1/_src/metadata/loading_test.py
@@ -12,15 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import dataclasses
+
 from absl.testing import absltest
 from etils import epath
 import jax
 import numpy as np
+from orbax.checkpoint import test_utils
 from orbax.checkpoint._src.metadata import value as value_metadata
-from orbax.checkpoint.experimental.v1._src.metadata import loading as metadata_loading
+import orbax.checkpoint.experimental.v1 as ocp
+from orbax.checkpoint.experimental.v1._src.context import context as context_lib
+from orbax.checkpoint.experimental.v1._src.context import options as options_lib
 from orbax.checkpoint.experimental.v1._src.metadata import types as metadata_types
-from orbax.checkpoint.experimental.v1._src.saving import saving
 from orbax.checkpoint.experimental.v1._src.testing import array_utils as array_test_utils
+from orbax.checkpoint.experimental.v1._src.testing import handler_utils
+
+
+Foo = handler_utils.Foo
+Bar = handler_utils.Bar
+AbstractFoo = handler_utils.AbstractFoo
+AbstractBar = handler_utils.AbstractBar
 
 
 class PyTreeMetadataTest(absltest.TestCase):
@@ -29,7 +40,7 @@ class PyTreeMetadataTest(absltest.TestCase):
     super().setUp()
     self.directory = epath.Path(self.create_tempdir().full_path) / 'ckpt'
     self.pytree, self.abstract_pytree = array_test_utils.create_numpy_pytree()
-    saving.save_pytree(self.directory, self.pytree)
+    ocp.save_pytree(self.directory, self.pytree)
 
   def _create_value_metadata(self, value):
     if isinstance(value, np.ndarray):
@@ -39,7 +50,7 @@ class PyTreeMetadataTest(absltest.TestCase):
       )
       return value_metadata.ArrayMetadata(
           name='a',
-          directory=self.directory,
+          directory=self.directory / 'pytree',
           shape=value.shape,
           dtype=value.dtype,
           sharding=None,
@@ -55,9 +66,9 @@ class PyTreeMetadataTest(absltest.TestCase):
       )
       return value_metadata.ScalarMetadata(
           name='a',
-          directory=self.directory,
+          directory=self.directory / 'pytree',
           shape=(),
-          dtype=dtype,
+          dtype=dtype,  # pytype: disable=wrong-arg-types
           sharding=None,
           storage=storage_metadata,
       )
@@ -66,18 +77,85 @@ class PyTreeMetadataTest(absltest.TestCase):
 
   def test_invalid_path(self):
     with self.assertRaises(FileNotFoundError):
-      metadata_loading.pytree_metadata(self.directory.parent)
+      ocp.pytree_metadata(self.directory.parent)
     with self.assertRaises(FileNotFoundError):
-      metadata_loading.pytree_metadata(self.directory.parent / 'foo')
+      ocp.pytree_metadata(self.directory.parent / 'foo')
 
   def test_pytree_metadata(self):
-    metadata = metadata_loading.pytree_metadata(self.directory)
+    metadata = ocp.pytree_metadata(self.directory)
     self.assertIsInstance(metadata, metadata_types.CheckpointMetadata)
-    self.assertIsInstance(metadata.metadata, metadata_types.PyTreeMetadata)
     expected_pytree_metadata = jax.tree.map(
         self._create_value_metadata, self.pytree
     )
-    self.assertEqual(expected_pytree_metadata, metadata.metadata.pytree)
+    self.assertEqual(expected_pytree_metadata, metadata.metadata)
+
+  def test_load_with_metadata(self):
+    metadata = ocp.pytree_metadata(self.directory)
+
+    def _set_numpy_cast_type(x):
+      if isinstance(x, np.ndarray):
+        return x.astype(np.int16)
+      elif isinstance(x, value_metadata.ArrayMetadata) and not isinstance(
+          x, value_metadata.ScalarMetadata
+      ):
+        return dataclasses.replace(x, dtype=np.int16)
+      else:
+        return x
+
+    metadata = dataclasses.replace(
+        metadata, metadata=jax.tree.map(_set_numpy_cast_type, metadata.metadata)
+    )
+    expected_pytree = jax.tree.map(_set_numpy_cast_type, self.pytree)
+
+    with self.subTest('pytree_metadata'):
+      loaded_pytree = ocp.load_pytree(self.directory, metadata.metadata)
+      test_utils.assert_tree_equal(self, expected_pytree, loaded_pytree)
+    with self.subTest('full_metadata'):
+      loaded_pytree = ocp.load_pytree(self.directory, metadata)
+      test_utils.assert_tree_equal(self, expected_pytree, loaded_pytree)
+
+
+
+class CheckpointablesMetadataTest(absltest.TestCase):
+
+  def setUp(self):
+    super().setUp()
+    self.directory = epath.Path(self.create_tempdir().full_path) / 'ckpt'
+    checkpointables_options = (
+        options_lib.CheckpointablesOptions.create_with_handlers(
+            handler_utils.FooHandler,
+            handler_utils.BarHandler,
+        )
+    )
+    self.enter_context(
+        context_lib.Context(checkpointables_options=checkpointables_options)
+    )
+    checkpointables = {
+        'foo': Foo(1, 'foo'),
+        'bar': Bar(2, 'bar'),
+    }
+    ocp.save_checkpointables(self.directory, checkpointables)
+
+  def test_checkpointables_metadata(self):
+    metadata = ocp.checkpointables_metadata(self.directory)
+    self.assertIsInstance(metadata, metadata_types.CheckpointMetadata)
+    self.assertIsInstance(metadata.metadata, dict)
+    self.assertIsInstance(metadata.metadata['foo'], AbstractFoo)
+    self.assertIsInstance(metadata.metadata['bar'], AbstractBar)
+
+  def test_load_with_metadata(self):
+    path1 = epath.Path(self.create_tempdir().full_path) / 'ckpt1'
+    path2 = epath.Path(self.create_tempdir().full_path) / 'ckpt2'
+
+    ocp.save_checkpointables(path1, {'foo': Foo(1, 'foo')})
+    ocp.save_checkpointables(path2, {'bar': Bar(3, 'bar')})
+
+    metadata = ocp.checkpointables_metadata(path1)
+    self.assertSameElements(metadata.metadata.keys(), ['foo'])
+    metadata = dataclasses.replace(metadata, metadata={'bar': AbstractFoo()})
+    loaded = ocp.load_checkpointables(path2, metadata)
+    self.assertSameElements(loaded.keys(), ['bar'])
+    self.assertEqual(Foo(3, 'bar'), loaded['bar'])