+++ checkpoint/orbax/checkpoint/experimental/v1/handlers.py
@@ -16,12 +16,16 @@
 
 # pylint: disable=g-importing-member, g-multiple-import, unused-import, g-bad-import-order
 
+import orbax.checkpoint.experimental.v1._src.handlers.global_registration
 from orbax.checkpoint.experimental.v1._src.handlers.types import (
     CheckpointableHandler,
 )
 
 from orbax.checkpoint.experimental.v1._src.handlers.pytree_handler import (
     PyTreeHandler,
+)
+from orbax.checkpoint.experimental.v1._src.handlers.proto_handler import (
+    ProtoHandler,
 )
 
 from orbax.checkpoint.experimental.v1._src.handlers.registration import (
@@ -30,5 +34,3 @@ from orbax.checkpoint.experimental.v1._src.handlers.registration import (
     local_registry,
     register_handler,
 )
-
-import orbax.checkpoint.experimental.v1._src.handlers.global_registration