+++ checkpoint/CHANGELOG.md
@@ -7,12 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- #v1 Add `checkpointables_metadata` function and simplify PyTree
+metadata presentation.
+
+## [0.11.11] - 2025-04-07
+
 ### Changed
 
 - Asynchronous directory creation enabled by default.
 - Remove `CheckpointInfo.is_locked` as it is not used anymore.
-
-## [0.11.11] - 2025-04-07
 
 ### Added